fix(sra): sync Epoch uint64 narrowing (cherry-picked 8c3eff9)
Apply wjmelements' upstream Epoch uint96->uint64 change and update every
downstream assumption:

- _qEnd range guard threshold: type(uint96).max -> type(uint64).max; the
  six uint96 wrap literals (POST_PERIOD / VERIFICATION_WINDOW / end) -> uint64
- MaxQuarter probes flip: at uint64 width, uint64.max x EPOCHS_PER_QUARTER
  >= 2^64 always overflows, so the guard (InvalidParameter) becomes the
  revert path for postVolume/correctVolume/finalizeConversion/submitShares
  (SRAAdversarial tests renamed *_RangeGuard_InvalidParameter)
- PendingTask storage layout shift: Epoch modified is now uint64, so the
  approvals bitmask sits at offset 64 instead of 96 — fix approvalsOf
  (SRAInvariant) accordingly (this was the I3a invariant regression)
- halmos harness/check wrap literals -> uint64

forge build / lint --deny notes / test: 14 suites, 296 passed, 0 failed

+++ src/ServiceRewardsActor.sol
@@ -201,50 +201,41 @@ contract ServiceRewardsActor is UnanimousGovernance {
     function _qEnd(uint64 q) internal view returns (Epoch) {
         // S1C: Q × EPOCHS_PER_QUARTER uses a uint256 intermediate to guard overflow.
         //
-        // Review-② range guard: Epoch is being narrowed upstream from uint96 to uint64 (f02
-        // consistency). Without an explicit check, an attacker-controlled huge q would wrap inside
-        // Epoch.wrap and could collide into the current quarter window, bypassing the window checks
-        // (enabling forged finalize/shares). The guard rejects end beyond the Epoch width.
-        //
-        // ⚠️ When the upstream Epoch PR narrows Epoch to uint64, sync ALL of the following:
-        //   1. this guard threshold: type(uint96).max -> type(uint64).max
-        //   2. the six `uint96` wrap literals in this file (Epoch.wrap(uint96(end)), and the
-        //      Epoch.wrap(uint96(POST_PERIOD)) / Epoch.wrap(uint96(VERIFICATION_WINDOW)) uses
-        //      in _inPostingWindow/_inVerificationWindow/_afterBinding/_frozenAtPostEnd) -> uint64
-        //   3. test/SRAAdversarial.t.sol: the four MaxQuarter tests flip back to expecting
-        //      InvalidParameter — at uint64 width, uint64.max × EPOCHS_PER_QUARTER ≥ 2^64 always
-        //      overflows, so the guard (not the window errors) becomes the revert path.
-        //      test_FinalizeConversion_MaxQuarter_RangeGuard (huge-EPOCHS_PER_QUARTER simulation)
-        //      remains valid as-is.
+        // Review-② range guard: Epoch narrowed upstream from uint96 to uint64 (f02 consistency,
+        // cherry-picked 8c3eff9). Without the explicit check, an attacker-controlled huge q would
+        // wrap inside Epoch.wrap and could collide into the current quarter window, bypassing the
+        // window checks (enabling forged finalize/shares). The guard rejects end beyond the Epoch
+        // width. At uint64 width, uint64.max × EPOCHS_PER_QUARTER ≥ 2^64 always overflows, so the
+        // guard is the revert path for the MaxQuarter probes (see test/SRAAdversarial.t.sol).
         uint256 end = uint256(ACTIVATION_EPOCH) + uint256(q) * uint256(EPOCHS_PER_QUARTER);
-        require(end <= type(uint96).max, InvalidParameter());
-        return Epoch.wrap(uint96(end));
+        require(end <= type(uint64).max, InvalidParameter());
+        return Epoch.wrap(uint64(end));
     }
 
     /// @dev posting window (E, E+POST].
     function _inPostingWindow(uint64 q) internal view returns (bool) {
         Epoch nowE = currentEpoch();
         Epoch e = _qEnd(q);
-        return nowE > e && nowE <= e + Epoch.wrap(uint96(POST_PERIOD));
+        return nowE > e && nowE <= e + Epoch.wrap(uint64(POST_PERIOD));
     }
 
     /// @dev verification window (E+POST, E+POST+VERIFY].
     function _inVerificationWindow(uint64 q) internal view returns (bool) {
         Epoch nowE = currentEpoch();
-        Epoch postEnd = _qEnd(q) + Epoch.wrap(uint96(POST_PERIOD));
-        return nowE > postEnd && nowE <= postEnd + Epoch.wrap(uint96(VERIFICATION_WINDOW));
+        Epoch postEnd = _qEnd(q) + Epoch.wrap(uint64(POST_PERIOD));
+        return nowE > postEnd && nowE <= postEnd + Epoch.wrap(uint64(VERIFICATION_WINDOW));
     }
 
     /// @dev post-binding: now > E+POST+VERIFY.
     function _afterBinding(uint64 q) internal view returns (bool) {
         Epoch nowE = currentEpoch();
-        Epoch verifyEnd = _qEnd(q) + Epoch.wrap(uint96(POST_PERIOD)) + Epoch.wrap(uint96(VERIFICATION_WINDOW));
+        Epoch verifyEnd = _qEnd(q) + Epoch.wrap(uint64(POST_PERIOD)) + Epoch.wrap(uint64(VERIFICATION_WINDOW));
         return nowE > verifyEnd;
     }
 
     /// @dev frozen determination at the E+POST instant (S5 snapshot semantics: derivable at any point, independent of call timing).
     function _frozenAtPostEnd(address orch, uint64 q) internal view returns (bool) {
-        return _isFrozenAt(orch, _qEnd(q) + Epoch.wrap(uint96(POST_PERIOD)));
+        return _isFrozenAt(orch, _qEnd(q) + Epoch.wrap(uint64(POST_PERIOD)));
     }
 
     /// @dev Determines whether the epoch falls inside any [freezeEpochs[i], unfreezeEpochs[i]) freeze interval.
